// client/src/agents/routes.tsx
export enum InsightRoutes {
  addUser = '/api/users/add',
  getAccountDetails = '/api/account/:accountId',
  getAccounts = '/api/accounts/:budgetId',
  getAllTransactions = '/api/transactions/all',
  getBudgets = '/api/budget',
  getBudgetYears = '/api/budget/years/:budgetId',
  getCategories = '/api/categories/:budgetId',
  getPayees = '/api/payees/:budgetId',
  getTransactionsByDay = '/api/transactions/:budgetId/aggregate',
  getTransactionsByPayee = '/api/transactions/:payeeId',
  reportsMonthly = '/api/reports/monthly',
  syncAccounts = '/api/sync/accounts',
  syncBudgets = '/api/sync/budgets',
  syncCategories = '/api/sync/categories',
  syncPayees = '/api/sync/payees',
  syncTransactions = '/api/sync/transactions',
  users = '/api/users/:budgetId',
}